+++ packages/discovery-react-components/src/components/SearchInput/__stories__/SearchInput.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, text, boolean, number, object } from '@storybook/addon-knobs/react';
+import { text, boolean, number, object } from '@storybook/addon-knobs/react';
 import SearchInput from '../SearchInput';
 import { SearchResults } from 'components/SearchResults/SearchResults';
 import { DiscoverySearch, DiscoverySearchProps } from 'components/DiscoverySearch/DiscoverySearch';
@@ -8,9 +8,6 @@ import { StoryWrapper, DummySearchClient } from 'utils/storybookUtils';
 import { createDummyResponsePromise } from 'utils/testingUtils';
 import DiscoveryV2 from 'ibm-watson/discovery/v2';
 import { action } from '@storybook/addon-actions';
-import defaultReadme from './default.md';
-import spellingSuggestionsReadme from './spellingSuggestions.md';
-import marked from 'marked';
 import { defaultMessages } from '../messages';
 
 const props = () => ({
@@ -65,46 +62,34 @@ const discoverySearchProps = (): DiscoverySearchProps => ({
 });
 
 storiesOf('SearchInput', module)
-  .addDecorator(withKnobs)
-  .add(
-    'default',
-    () => {
-      autocompletions = generateCompletionsArray(props().completionsCount, '');
-      return (
-        <StoryWrapper>
-          <DiscoverySearch {...discoverySearchProps()}>
-            <SearchInput {...props()} />
-          </DiscoverySearch>
-        </StoryWrapper>
-      );
-    },
-    {
-      info: marked(defaultReadme)
-    }
-  )
-  .add(
-    'with spelling suggestions',
-    () => {
-      const spellingSuggestionProps = {
-        ...discoverySearchProps(),
-        overrideQueryParameters: {
-          naturalLanguageQuery: 'Waston'
-        },
-        overrideSearchResults: {
-          suggested_query: 'Watson'
-        }
-      };
+  .addParameters({ component: SearchInput })
+  .add('default', () => {
+    autocompletions = generateCompletionsArray(props().completionsCount, '');
+    return (
+      <StoryWrapper>
+        <DiscoverySearch {...discoverySearchProps()}>
+          <SearchInput {...props()} />
+        </DiscoverySearch>
+      </StoryWrapper>
+    );
+  })
+  .add('with spelling suggestions', () => {
+    const spellingSuggestionProps = {
+      ...discoverySearchProps(),
+      overrideQueryParameters: {
+        naturalLanguageQuery: 'Waston'
+      },
+      overrideSearchResults: {
+        suggested_query: 'Watson'
+      }
+    };
 
-      return (
-        <StoryWrapper>
-          <DiscoverySearch {...spellingSuggestionProps}>
-            <SearchInput {...props()} />
-            <SearchResults />
-          </DiscoverySearch>
-        </StoryWrapper>
-      );
-    },
-    {
-      info: marked(spellingSuggestionsReadme)
-    }
-  );
+    return (
+      <StoryWrapper>
+        <DiscoverySearch {...spellingSuggestionProps}>
+          <SearchInput {...props()} />
+          <SearchResults />
+        </DiscoverySearch>
+      </StoryWrapper>
+    );
+  });