+++ examples/discovery-search-app/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2-beta.7](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.2-beta.5...discovery-search-app@1.0.2-beta.7) (2019-12-19)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.2-beta.6](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.2-beta.5...discovery-search-app@1.0.2-beta.6) (2019-12-19)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.2-beta.5](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.2-beta.3...discovery-search-app@1.0.2-beta.5) (2019-12-19)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
 ## [1.0.2-beta.4](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.2-beta.3...discovery-search-app@1.0.2-beta.4) (2019-12-11)
 
 **Note:** Version bump only for package discovery-search-app